New "worker" images
+ dt
+ locus_plot
+ annot_plot

These 3 have to do with the visualization of the snp2gene results and should be removed at some point in the feature because are (not that much) computationally expensive operations. It can be done better using pure php. There are some commented out approaches in the corresponding controllers that can be modified and used very easily.