php-fpm
Increase max_input_vars to 5000 from the default value.
The reason we did that is because during the submission of snp2gene jobs if a user selects multiple datasets or "all" options the number of variables that are being sent to the server is more than 1000 and the server starts truncating the vars resulting to errors.